+++ src/macro.jl
@@ -1,23 +1,52 @@
+
 """
-    @submit! [option=value]... expr
     @submit [option=value]... expr
 
 Submit a job from `expr`. If a `Job` is **explicitly** shown in `expr`, `DONE => job` will be automatically added to the dependency list. 
 
-- `option = value`: kwargs of [`Job`](@ref).
+- `expr`: any type of `Expr`ession is supported. 
 
-!!! warn "Only explicit jobs can be automatically added to dependency"
-    If a job is in a container, `@submit!` cannot know the elements in the container:
+- `option = value`: kwargs of [`Job`](@ref). If `expr` is parsed to be a `Pipelines.Program`, `option`s also include its inputs, outputs and run kwargs.
+
+See also [`Job`](@ref), [`submit!`](@ref)
+
+## Example
+
+```julia
+j = @submit 1+1
+wait(j)
+@assert result(j) == 2
+
+# you can use any keyword arguments that `Job` supports, such as `name`, `ncpu`:
+j_2sec = @submit name = "run after 2 sec" begin sleep(2); 32 end
+
+# because `j_2sec isa Job`, `DONE => j_2sec` is pushed to `j2.dependency`, and the `j_2sec` in the begin-end block is converted to `result(j_2sec)`:
+j2 = @submit mem=2KB begin
+    1 + result(j_2sec)
+end
+
+wait(j2)
+@assert result(j2) == 1 + 32
+
+# you can also manually add dependencies that not in the `expr`:
+j3 = @submit dependency = [PAST => j] println("j3 finished. result of j2 = ", result(j2))
+
+# Note: j3.dependency might be empty after submit, because JobScheduler will remove jobs that reached their states in the dependency list.
+```
+
+
+!!! warn "Known Issues: Only explicit jobs can be automatically added to dependency"
+    If a job is in a container, `@submit` cannot know the elements in the container:
     ```julia
-    j = @submit! 666
+    j = @submit 666
 
     jobs = Job[]  # the job container
     for i in 1:10
-        push!(jobs, @submit! begin sleep(2);i end) # 10 jobs will be added to `jobs`
+        push!(jobs, @submit begin sleep(30);i end) # 10 jobs will be added to `jobs`
     end
 
     x = 0
-    j_something_wrong = @submit! for j in jobs
+    j_something_wrong = @submit for j in jobs
         # have to use global x
         global x += result(j)
     end
@@ -27,6 +56,9 @@ Submit a job from `expr`. If a `Job` is **explicitly** shown in `expr`, `DONE =>
     result(j_something_wrong)
     # MethodError(+, (nothing, nothing), 0x0000000000007b16)
 
+    # To avoid it, we can 
+    #      (1) use `submit!`, or
+    #      (2) explicitly add `dependency = jobs` to `@submit`.
     x = 0
     j_ok = submit!(dependency = jobs) do
         for j in jobs
@@ -38,7 +70,7 @@ Submit a job from `expr`. If a `Job` is **explicitly** shown in `expr`, `DONE =>
     @assert x == 55
 
     x = 100
-    j_ok_too = @submit! dependency = jobs for j in jobs
+    j_ok_too = @submit dependency = jobs for j in jobs
         # have to use global x
         global x += result(j)
     end
@@ -46,61 +78,72 @@ Submit a job from `expr`. If a `Job` is **explicitly** shown in `expr`, `DONE =>
     @assert x == 155
     ```
 
-See also [`Job`](@ref), [`submit!`](@ref)
-
-## Example
-
-```julia
-j = @submit! 1+1
-wait(j)
-@assert result(j) == 2
-
-# you can use any keyword arguments that `Job` supports, such as `name`, `ncpu`:
-j_2sec = @submit! name = "run after 2 sec" begin sleep(2); 32 end
-
-# because `j_2sec isa Job`, `DONE => j_2sec` is pushed to `j2.dependency`, and the `j_2sec` in the begin-end block is converted to `result(j_2sec)`:
-j2 = @submit! mem=2KB begin
-    1 + result(j_2sec)
-end
-
-wait(j2)
-@assert result(j2) == 1 + 32
-
-# you can also manually add dependencies that not in the `expr`:
-j3 = @submit! dependency = [PAST => j] println("j3 finished. result of j2 = ", result(j2))
-
-# Note: j3.dependency might be empty after submit, because JobScheduler will remove jobs that reached their states in the dependency list.
-```
 """
-macro submit!(args...)
+macro submit(args...)
     local opts = args[1:end-1]
     local expr = args[end]
     
     # construct Job(...)
-    local params = opt2parameters(opts) # can esc
+    local params = opt2parameters(opts)
 
     # construct new_deps = [], and add deps in new_deps
     local possible_deps = sym_list(expr)  # Symbol[all symbols]
-    unique!(possible_deps)
     local dep_struct = Expr(:ref, :Any, possible_deps...) # Any[possible_deps...]
 
-    return quote
-        local job = Job($(esc(params)), () -> $(esc(expr)))
-        local deps = $(esc(dep_struct))
-        filter!(isajob, deps)
+    # if expr is only a Symbol: it could be a Program
+    local expr_is_a_symbol = expr isa Symbol
 
-        if !isempty(deps)
-            for dep in deps
-                push!(job.dependency, :done => dep)
-            end            
+    return quote
+        local job
+        if $expr_is_a_symbol  # could be a program
+            local evaluated = $(esc(expr))
+            if evaluated isa Program
+                job = Job($(esc(params)), evaluated)
+                @goto submit
+            else
+                @goto normal
+            end
+        else
+            @label normal
+            local deps = $(esc(dep_struct))
+            filter!(isajob, deps)
+            job = Job($(esc(params)), () -> $(esc(expr)))
+
+            if !isempty(deps)
+                for dep in deps
+                    push!(job.dependency, :done => dep)
+                end
+            end
+            @label submit
+            submit!(job)
         end
-        @info job.dependency
-        submit!(job)
     end
 end
 
+# function job_to_result!(expr::Expr, dep_syms::Vector{Symbol})
+#     for (i,a) in enumerate(expr.args)
+#         @inbounds expr.args[i] = job_to_result!(a, dep_syms)
+#     end
+#     expr
+# end
+# function job_to_result!(expr::Symbol, dep_syms::Vector{Symbol})
+#     if expr in dep_syms
+#         :(JobSchedulers.result_or_self($expr))
+#     else
+#         expr
+#     end
+# end
+# function job_to_result!(expr, dep_syms::Vector{Symbol})
+#     expr
+# end
+
 isajob(x::Job) = true
 isajob(x) = false
+isnotajob(x::Job) = false
+isnotajob(x) = true
+
+# result_or_self(x::Job) = result(x)
+# result_or_self(x) = result(x)
 
 function opt2parameters(opts::NTuple{N, Expr}) where N
     for opt in opts
@@ -114,20 +157,34 @@ end
 """
     sym_list(x)
 
-Extract var (symbols) in `var isa Job`.
+Extract var (symbols) in expression/symbol.
 """
 function sym_list(x)
     v = Vector{Symbol}()
-    sym_list!(v, x)
+    just_defined = Set()
+    _sym_list!(v, just_defined, x)
+    unique!(v)
+    filter!(v) do x
+        !(x in just_defined)
+    end
     v
 end
 
-sym_list!(v::Vector, expr::Expr) = sym_list!(v, expr.args)
-sym_list!(v::Vector, sym::Symbol) = push!(v, sym)
-sym_list!(v::Vector, x) = nothing
+function _sym_list!(v::Vector, just_defined::Set, expr::Expr)
+    if expr.head == :(=) && length(expr.args) == 2
+        push!(just_defined, expr.args[1])
+        # skip first arg
+        _sym_list!(v, just_defined, expr.args[2])
+    else
+        _sym_list!(v, just_defined, expr.args)
+    end
+    expr
+end
+_sym_list!(v::Vector, just_defined::Set, sym::Symbol) = push!(v, sym)
+_sym_list!(v::Vector, just_defined::Set, x) = nothing
 
-function sym_list!(v::Vector, args::Vector)
+function _sym_list!(v::Vector, just_defined::Set, args::Vector)
     for arg in args
-        sym_list!(v::Vector, arg)
+        _sym_list!(v::Vector, just_defined, arg)
     end
 end