+++ test/test_macro.jl
@@ -1,19 +1,19 @@
 
 
-@testset "@submit!" begin
-    j = @submit! name = "abc" dependency = [] 1+1
+@testset "@submit" begin
+    j = @submit name = "abc" dependency = [] 1+1
     wait(j)
     @test result(j) == 2
     @test j.name == "abc"
 
     a = 3242
-    j = @submit! a + 5
+    j = @submit a + 5
     wait(j)
     @test result(j) == 3242 + 5
 
-    j_long = @submit! name = "abc" dependency = j begin sleep(2); 32 end
+    j_long = @submit name = "abc" dependency = j begin sleep(2); 32 end
 
-    j2 = @submit! mem = 2KB begin 1 + result(j) + result(j_long) end
+    j2 = @submit mem = 2KB begin 1 + result(j) + result(j_long) end
     @test length(j2.dependency) > 0
     sleep(1)
     @test j2.state === QUEUING
@@ -23,7 +23,7 @@
     function experiments_jobschedulers2(a, K=10000)
         x = 0
         for i in 1:K
-            @submit! x += a
+            @submit x += a
         end
         wait_queue()
         x
@@ -35,18 +35,18 @@
         y = [a:a+K...]
         js = Job[]
         
-        # for _ in 1:K
-            j1 = @submit! sum(y)/length(y)
+        for _ in 1:K
+            j1 = @submit sum(y)/length(y)
             for i in y
-                j2 = @submit! (i - result(j1))^2
+                j2 = @submit (i - result(j1))^2
                 push!(js, j2)
             end
-        # end
-        jsum = @submit! for j in js
+        end
+        jsum = @submit dependency=js for j in js
             x += result(j)
         end
         wait(jsum)
-        @error "jsum's dependency is outer j, which does not make sense!"
-        x
+        x/K
     end
+    @test_nowarn test_sequential(4,50)
 end